% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Coagulation.R
\name{coagulation}
\alias{coagulation}
\title{Coagulation Data File}
\description{
A data set of 24 animals coagualtion collected in 4 random groups and the samples were geenrated randomly
}
\details{
\itemize{
  \item coag. Blood Coagulation
  \item diet. Random Diet Group
  ...
}
}

