% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/oneway.R
\name{oneway factor function}
\alias{oneway factor function}
\alias{oneway.factor}
\title{Factor of Anova}
\usage{
\method{oneway}{factor}(z, y, ...)
}
\arguments{
\item{z}{A vector that contains groups that have factors}

\item{y}{A vector that contains the numerical values}
}
\value{
oneway.factor  A list
}
\description{
Factor of Anova
}

