% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/oneway.R
\name{oneway formula function}
\alias{oneway formula function}
\alias{oneway.formula}
\title{oneway formula of Anova}
\usage{
\method{oneway}{formula}(formula, data = list(), ...)
}
\arguments{
\item{formula}{A formula}

\item{data}{A list that contains data}
}
\value{
oneway.factor  A list
}
\description{
oneway formula of Anova
}

