% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/oneway.R
\name{oneway lsmeans function}
\alias{lsmeans.oneway}
\alias{oneway lsmeans function}
\title{LS Means of Anova}
\usage{
lsmeans.oneway(object, ...)
}
\arguments{
\item{x}{A object}
}
\value{
lsmeans.oneway  A table of Fisher's LSD results
}
\description{
LS Means of Anova
}

