% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/oneway.R
\name{oneway print function}
\alias{oneway print function}
\alias{print.oneway}
\title{Print oneway of Anova}
\usage{
\method{print}{oneway}(x, ...)
}
\arguments{
\item{x}{A list}
}
\value{
print.oneway  A short table anova.
}
\description{
Print oneway of Anova
}

