% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/oneway.R
\name{oneway Print Summary function}
\alias{oneway Print Summary function}
\alias{print.summary.oneway}
\title{Print Summary of Anova}
\usage{
\method{print}{summary.oneway}(x, ...)
}
\arguments{
\item{x}{A object}
}
\value{
print.summary.oneway  A short table of ANOVA that contains , F-Test data, MSW, MSB, and P value
}
\description{
Print Summary of Anova
}

