% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/oneway.R
\name{oneway summary function}
\alias{oneway summary function}
\alias{summary.oneway}
\title{Summary of Anova}
\usage{
\method{summary}{oneway}(object, ...)
}
\arguments{
\item{x}{A object}
}
\value{
summary.oneway  A short table of ANOVA that contains , F-Test data, MSW, MSB, and P value
}
\description{
Summary of Anova
}

