% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/oneway.R
\name{oneway}
\alias{oneway}
\alias{oneway.default}
\title{Thilanks Anova}
\usage{
oneway(z, ...)

\method{oneway}{default}(z, ...)
}
\arguments{
\item{z}{A list that contains groups that have factors}
}
\value{
oneway  A list that contains the Sum of Squares Between (ssb), Sum of Squares Within (ssw), Degrees of Freedom Between (dfb), Degrees of Freedom Wining (dfw), Group Mean (GroupMean), Names of the Groups (namesGroups)
}
\description{
Thilanks Anova
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: oneway.default(x, ...)
}}

